- Change data type of `miss` and `match` to i8 - Change data type of `gap_open` and `gap_extend` to u8 - Change data type of score which in `blosum50`, `blosum62` and `pam120` to i8

+++ src/cli.rs
@@ -13,25 +13,25 @@ pub struct CliArgs
     #[clap(default_value_t = 1)]
     #[clap(value_parser = is_integer)]
     #[clap(help = "Penalty score when two residue missmatch")]
-    pub _miss: i16,
+    pub _miss: i8,
 
     #[clap(name = "match", long, short = 'm', display_order = 2)]
     #[clap(default_value_t = 1)]
     #[clap(value_parser = is_integer)]
     #[clap(help = "Add this score when two residue match")]
-    pub _match: i16,
+    pub _match: i8,
     
     #[clap(name = "gap-open", long, short = 'o', display_order = 3)]
     #[clap(default_value_t = 3)]
     #[clap(value_parser = is_positive_integer)]
     #[clap(help = "Gap open penalty score, positive integer")]
-    pub gap_open: u16,
+    pub gap_open: u8,
 
     #[clap(name = "gap-extend", long, short = 'e', display_order = 4)]
     #[clap(default_value_t = 2)]
     #[clap(value_parser = is_positive_integer)]
     #[clap(help = "Gap extend penalty score, positive integer")]
-    pub gap_extend: u16,
+    pub gap_extend: u8,
 
     #[clap(name = "protein", long, short = 'p', display_order = 5)]
     #[clap(conflicts_with_all = &["miss", "match"])]
@@ -54,13 +54,13 @@ pub struct CliArgs
     pub query: String
 }
 
-fn is_integer(arg: &str) -> Result<i16, String>
+fn is_integer(arg: &str) -> Result<i8, String>
 {
     let score = arg.parse().map_err(|_expection| format!("{} not a i8 integer", arg))?;
     Ok(score)
 }
 
-fn is_positive_integer(arg: &str) -> Result<u16, String>
+fn is_positive_integer(arg: &str) -> Result<u8, String>
 {
     let score = arg.parse().map_err(|_expection| format!("{} not a non-negative integer", arg))?;
     Ok(score)
